refactor: adjusting architecture and user type
