+++ src/api/modules/user/user.input.ts
@@ -1,7 +1,7 @@
 import { InputType, Field } from 'type-graphql';
 
 @InputType()
-export class CreateUserInput {
+export class UserInput {
   @Field()
   name!: string;
