+++ src/api/schema.ts
@@ -1,8 +1,8 @@
 import 'graphql-import-node';
 import { GraphQLSchema } from 'graphql';
 import { buildSchemaSync } from 'type-graphql';
-import { HelloResolver } from './resolver/hello-resolver';
-import { UserResolver } from './resolver/user-resolver';
+import { HelloResolver } from './modules/hello/hello.resolver';
+import { UserResolver } from './modules/user/user.resolver';
 
 // Load resolvers
 export const Schema: GraphQLSchema = buildSchemaSync({
