+++ src/data/entity/user.entity.ts
@@ -1,0 +1,28 @@
+import { Entity, BaseEntity, PrimaryGeneratedColumn, Column, CreateDateColumn, UpdateDateColumn } from 'typeorm';
+
+@Entity('user')
+export class UserEntity extends BaseEntity {
+  @PrimaryGeneratedColumn('uuid')
+  id!: number;
+
+  @Column()
+  name!: string;
+
+  @Column({ unique: true })
+  email!: string;
+
+  @Column()
+  password!: string;
+
+  @Column({ type: 'date' })
+  birthDate!: Date;
+
+  @Column({ default: true })
+  isActive!: boolean;
+
+  @CreateDateColumn()
+  creationDate!: Date;
+
+  @UpdateDateColumn()
+  updateDate!: Date;
+}
